run course sync in parallel
